fz.tint404011_move_to_color_temp
