+++ src/converters/fromZigbee.ts
@@ -1989,40 +1989,6 @@ export const tint_scene: Fz.Converter<"genBasic", undefined, "write"> = {
         return payload;
     },
 };
-export const tint404011_move_to_color_temp: Fz.Converter<"lightingColorCtrl", undefined, "commandMoveToColorTemp"> = {
-    cluster: "lightingColorCtrl",
-    type: "commandMoveToColorTemp",
-    convert: (model, msg, publish, options, meta) => {
-        // The remote has an internal state so store the last action in order to
-        // determine the direction of the color temperature change.
-        if (!globalStore.hasValue(msg.endpoint, "last_color_temp")) {
-            globalStore.putValue(msg.endpoint, "last_color_temp", msg.data.colortemp);
-        }
-
-        const lastTemp = globalStore.getValue(msg.endpoint, "last_color_temp");
-        globalStore.putValue(msg.endpoint, "last_color_temp", msg.data.colortemp);
-        let direction = "down";
-        if (lastTemp > msg.data.colortemp) {
-            direction = "up";
-        } else if (lastTemp < msg.data.colortemp) {
-            direction = "down";
-        } else if (msg.data.colortemp === 370 || msg.data.colortemp === 555) {
-            // The remote goes up to 370 in steps and emits 555 on down button hold.
-            direction = "down";
-        } else if (msg.data.colortemp === 153) {
-            direction = "up";
-        }
-
-        const payload = {
-            action: postfixWithEndpointName("color_temperature_move", msg, model, meta),
-            action_color_temperature: msg.data.colortemp,
-            action_transition_time: msg.data.transtime,
-            action_color_temperature_direction: direction,
-        };
-        addActionGroup(payload, msg, model);
-        return payload;
-    },
-};
 export const ewelink_action: Fz.Converter<"genOnOff", undefined, ["commandOn", "commandOff", "commandToggle"]> = {
     cluster: "genOnOff",
     type: ["commandOn", "commandOff", "commandToggle"],
